pyproject.toml
[tool.poetry]
name = "esg_llm"
version = "0.1.0"
description = ""
authors = ["DoctorSlimm <87606888+DoctorSlimm@users.noreply.github.com>"]
readme = "README.md"
packages = [
    { include = "esg_llm" },
]

[tool.poetry.dependencies]
python = "^3.9"
ninja = "*"
psutil = "*"
ray = ">=2.5.1"
pandas = "*"
pyarrow = "*"
sentencepiece = "*"
numpy = "*"
einops = "*"
transformers = ">=4.34.0"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
pydantic = "1.10.13"
aioprometheus = {extras = ["starlette"], version = "*"}
python-dotenv = "*"
openai = "^1.3.4"
abc123abc = "^0.0.0"
modal = "^0.55.4155"
langchain = "^0.0.345"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
